* Fixing error with executeshell

+++ cyEngine/cyUtilities/source/cyArgumentParser.cpp
@@ -8,7 +8,10 @@ ArgumentParser::parse(const String& parameters) {
   for (auto argument : m_argumentMap) {
     if (parameters.find(argument.first) != String::npos) {
       String tmpString = Utils::BLANK;
-      for (int32 i = static_cast<int32>(parameters.find(argument.first)) + argument.first.length() + 1; i < parameters.length(); ++i) {
+      for (int32 i = static_cast<int32>(parameters.find(argument.first)) + 
+             static_cast<int32>(argument.first.length() + 1); 
+           i < static_cast<int32>(parameters.length()); 
+           ++i) {
         if (parameters[i] == '-') { break; }
         tmpString += parameters[i];
       }
